% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_suppl_information.R
\name{add_suppl_information}
\alias{add_suppl_information}
\title{Add supplemental information column(s) to the result table}
\usage{
add_suppl_information(table, info)
}
\arguments{
\item{table}{table with statistical results with protein or gene ID as row.names}

\item{info}{.txt file with at least two columns and with an header. first column correspond to gene of protein ID
other columns correspond to interest information E}
}
\value{
Combine \code{table} and \code{info} table
}
\description{
\code{add_suppl_information} generates a new table with new information in supplemental columns
}
