% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBiomaRtAnnotation.R
\name{addBiomaRtAnnotation}
\alias{addBiomaRtAnnotation}
\title{Automatic annotation using BiomartR package}
\usage{
addBiomaRtAnnotation(
  normObj,
  biomart = "plants_mart",
  dataset = "athaliana_eg_gene",
  host = "plants.ensembl.org",
  features = c("ensembl_peptide_id", "ensembl_transcript_id", "ensembl_gene_id")
)
}
\arguments{
\item{normObj}{dataframe with row.names that contains gene, protein or transcript ID}

\item{biomart}{\code{plants_mart} by default (for plants). For other organisms, use  "ENSEMBL_MART_ENSEMBL",
see \code{listMarts} biomartR function}

\item{dataset}{\code{"athaliana_eg_gene"} by default. biomartR dataset that corresponds to the studied organism.
List of available datasets can be obtained using  biomartR \code{listDatasets} function}

\item{host}{\code{"www.plants.ensembl.org"} by default. Can be "www.ensembl.org" to get other dataset.}

\item{features}{\code{"ensembl_peptide_id"}, \code{"ensembl_transcript_id"} ou \code{"ensembl_gene_id"}.
\code{"ensembl_peptide_id"} by default. To be adjusted according to ID in the row.names of the table}
}
\value{
Return a result table with additional columns that contain gene or protein annotation.
}
\description{
\code{addBiomaRtAnnotation} is a function to retrieve protein, transcript or gene annotation
}
