% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_edgeR.R
\name{batchVerification}
\alias{batchVerification}
\title{Statistical model from the EdgeR package
\code{batchVerification} Function to check that batch was indicated in the SampleTable}
\usage{
batchVerification(IPObj)
}
\arguments{
\item{IPObj}{object, that comes from the function \code{loadData}}
}
\value{
Return \code{TRUE} if column with batch information exists
}
\description{
Statistical model from the EdgeR package
\code{batchVerification} Function to check that batch was indicated in the SampleTable
}
