% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTable.R
\name{createTable}
\alias{createTable}
\title{Interactive table}
\usage{
createTable(tres)
}
\arguments{
\item{tres}{Table containing an header and one name for each lines}
}
\value{
Return interactive table in html format
}
\description{
\code{createTable} create an html interactive table using DT package
}
\details{
Interactive table containing a search option and allowing to select and export slected data using csv and xls.
}
