% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_edgeR.R
\name{glm_edgeR}
\alias{glm_edgeR}
\title{Genewise negative binomial generalized linear model developped by the EdgeR}
\usage{
glm_edgeR(
  IPObj,
  control,
  treatment = NULL,
  batchE = FALSE,
  div = NULL,
  fnm = NULL,
  filter = NULL,
  min.disp = NULL,
  abundance.trend = TRUE,
  glm = NULL
)
}
\arguments{
\item{IPObj}{object from function \code{loadData}.}

\item{control}{Name of the control. The name must be concordant with one of the conditions' name of the sample
table (2nd column)}

\item{treatment}{\code{NULL} by default, will take the name of the second condition in the level order. Can be
changed by a name included in the conditions column of the sample table}

\item{batchE}{\code{FALSE} by default. If there is batch effect in the third column of the sample table, can be set to \code{TRUE}}

\item{div}{Scale factor}

\item{fnm}{\code{NULL} or a character string with the treatment factor name,
as used in the column names of the factors data frame, and in the formula.}

\item{filter}{Minimal number of spectra}

\item{min.disp}{Minimal number to calculate dispersion}

\item{abundance.trend}{argument of EdgeR function}

\item{glm}{Genewise Negative Binomial Generalized Linear Model with \code{"QL"} or without \code{"classic"} Quasi-likelihood tests.}
}
\value{
Return tables with four column
}
\description{
\code{modifFINALEdgeR} Function
}
