% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlPlot.R
\name{htmlPlot}
\alias{htmlPlot}
\title{Interactive volcanoplot using the R package Plotly}
\usage{
htmlPlot(
  tres,
  sign = c("adjp", "p.value"),
  max.pval = 0.05,
  min.LFC = 1,
  listGenes = c(),
  colforcolor = NULL,
  custom_text = NULL
)
}
\arguments{
\item{tres}{Table obtained from the \code{stat_test} function, that is one table with one ID by row and
at least one column with \code{logFC} and one other with  \code{adjp}}

\item{sign}{Choice a the value used to show statistical significance: \code{adjp} (default) is \code{p.value}}

\item{max.pval}{Cut-off of p-value considered as relevant}

\item{min.LFC}{Minimum LogFC for a protein to be considered as significant}

\item{listGenes}{Optional, list of gene/protein ID to be colored}

\item{colforcolor}{Optional, colored protein according to information in a supplemental table}

\item{custom_text}{Optional, add text labels}
}
\value{
Return html interactive graph
}
\description{
\code{htmlPlot} creates an interactive volcanoplot showing statistical significance (p-value or adjusted p-value) vs protein enrichment (LogFC)
}
\details{
If \code{listGenes} is empty, protein with significantly different FC will be highlighted
If \code{listGenes}, is filled, only proteins contained in the list will be filled
}
