% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_pheatmap.R
\name{IP_pheatmap}
\alias{IP_pheatmap}
\title{Heatmap using the package pheatmap}
\usage{
IP_pheatmap(
  IPObj,
  GeneList,
  norm = c("nothing", "total", "DEseq"),
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  annotation_col = NULL,
  annotation_row = NULL,
  fontsize_row = 6,
  title = NULL
)
}
\arguments{
\item{IPObj}{Object generated from the function \code{loadData}}

\item{GeneList}{List of selected genes/proteins used to plot for the heatmap.}

\item{norm}{If \code{"nothing"}, no normalization, if  \code{"total"}, counts are divided by the total number of counts,
if \code{"DEseq"}, counts are normalized using the median of ratios (DEseq method)}

\item{cluster_rows}{Determine if rows should be clustered, it can be \code{TRUE} or \code{FALSE}}

\item{cluster_cols}{Determine if columns should be clustered, it can be \code{TRUE} or \code{FALSE}}

\item{annotation_col}{Data frame that specifies the annotations shown on left side of the heatmap.}

\item{annotation_row}{Data frame that specifies the annotations shown on top of the heatmap.}

\item{fontsize_row}{To change font size of the row label. By defaut, the font size is 6}

\item{title}{To change graph title}
}
\value{
Return heatmap for selected genes/proteins
}
\description{
\code{IP_pheatmap} uses the pheatmap package to plot heatmaps showing spectral counts across samples for selected genes/proteins.
}
\details{
This function extract count data for selected genes \code{GeneList} and create an heatmap for all IP in the \code{IPObj}
}
