% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.unique.R
\name{is.unique}
\alias{is.unique}
\title{Create sub dataset
Timothee Vincent
Check that a vector is unique}
\usage{
is.unique(vect)
}
\arguments{
\item{vect}{a vector}
}
\value{
return \code{TRUE} if the vector contains only unique elements, return \code{FALSE} otherwise
}
\description{
\code{is.unique} use vector as argument and check that all vector elements are unique.
If true, return \code{TRUE}, otherwise return \code{FALSE}
}
