% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_IP_Data.R
\name{load_IP_Data}
\alias{load_IP_Data}
\title{Load spectral count data and sample description.}
\usage{
load_IP_Data(countFile, sampleTable)
}
\arguments{
\item{countFile}{tab delimited file (.txt) that contain all spectral counts with one distinct column for each IP experiments.
The first column of the files should contains protein names. The file needs to contain one header.}

\item{sampleTable}{tab delimited file (.txt) with two or three columns and an header. The first column correspond to the experiment names.
Each name should be unique and correspond to the names indicated in the header of \code{exprsFile}.
The second column indicates the different conditions (treatment, mutation). The third column is optional
and is used to indicate a batch effect.}
}
\value{
\code{IPObj} object type that contains the dataframe with spectral counts and information about treatment and eventual batch effect
}
\description{
Function to load spectral count data. Spectral count for all IPs should be stored in one
tab delimited file with one distinct column for each IP experiments. The first column of the files should contains protein names.
The file needs to contain one header but must not contain any special characters. Name must not start by numbers.
}
\examples{
CountTable <- system.file("extdata", "CountTable.txt", package = "IPinquiry4")
SampleTable <- system.file("extdata", "SampleTable.txt", package = "IPinquiry4")
IP_data <- load_IP_Data(CountTable, SampleTable)
}
