% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSplot.R
\name{MDSplot}
\alias{MDSplot}
\title{Create Multidimensional scaling (MDS) plots}
\usage{
MDSplot(IPObj, norm = c("nothing", "total", "DEseq"))
}
\arguments{
\item{IPObj}{object that comes directly from the function \code{load_IP_Data}}

\item{norm}{(default nothing), if \code{norm}="total", all counts are divided by the total number of count,
if \code{norm}="DEseq", counts are normalized according to the median-to-ratios methods as used in the DEseq2 R package}
}
\value{
Multidimensional scaling (MDS) plot based on euclidean distances
}
\description{
Function to plot data in a 2D plan such that the between-object distances are preserved as well as possible.
\code{MDSplot} is  useful to visualise the level of similarity between experiments
}
