% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_of_ratios.R
\name{median_of_ratios}
\alias{median_of_ratios}
\title{Function to calculate the scale factor as done by DEseq package (Median-to-ratio method)}
\usage{
median_of_ratios(data)
}
\arguments{
\item{data}{Objet de type IPObj, provenant directement de la fonction \code{loadData}.}
}
\value{
Return the scale factor calculated using the median to ratio method as used in DEseq2
}
\description{
Function to calculate the scale factor using the median to ratio method as used in DEseq2
}
