% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHmakeRandomString.R
\name{MHmakeRandomString}
\alias{MHmakeRandomString}
\title{Create sub dataset
Timothee Vincent
Generated 12 letter random string}
\usage{
MHmakeRandomString(n = 1, length = 12)
}
\arguments{
\item{n}{(default 1). Number of generated randomstring}

\item{length}{(default 12), Size of the random string}
}
\value{
Return a random string
}
\description{
\code{MHmakeRandomString} Annexe function called by the function \code{moreThantwoCond}. It is used to create temprory file.
This file will be erased at the end
}
