% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moreThantwoCond.R
\name{moreThantwoCond}
\alias{moreThantwoCond}
\title{Create sub dataset}
\usage{
moreThantwoCond(IPObj, cond1, cond2, batchE = FALSE)
}
\arguments{
\item{IPObj}{Object loaded by the loadData function}

\item{cond1}{Name of the first condition (control)}

\item{cond2}{Name of the second condition}

\item{batchE}{Determine if a batch effect in taken in the statistical model}
}
\value{
Return an IPObj type object that contains only the count for the selected conditions
}
\description{
\code{moreThantwoCond} function called by the statistical test function (\code{modifFINALEdgeR})
to reduce the dataset to two conditions when more are in the input table.
}
