% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDF_Plot.R
\name{PDF_Plot}
\alias{PDF_Plot}
\title{Create volcanoplot graph suitable for pdf export}
\usage{
PDF_Plot(
  tres,
  sign = c("adjp", "p.value"),
  max.pval = 0.05,
  min.LFC = 1,
  line = TRUE,
  point_color = NULL,
  min_x = NULL,
  max_x = NULL,
  min_y = NULL,
  max_y = NULL,
  colforcolor = NULL,
  point_size = 4,
  label = TRUE,
  label_size = 4,
  custom_text = NULL,
  title = NULL
)
}
\arguments{
\item{tres}{table obtained from the  \code{stat_test} function, that is one table with one ID by row and
at least one column with \code{logFC} and one other with  \code{adjp} or \code{p.value}}

\item{sign}{Choice a the value used to show statistical significance: \code{adjp} (default) is \code{p.value}}

\item{max.pval}{Cut-off p-value considered as relevant}

\item{min.LFC}{Minimum LogFC for a protein to be considered as significant}

\item{line}{Dotted red lines to show FC and p-value cut-off. it can be \code{TRUE} or \code{FALSE}}

\item{point_color}{Optional, allow for manually setting colors}

\item{min_x}{Optional, set min value for x axis}

\item{max_x}{Optional, set max value for x axis}

\item{min_y}{Optional, set min value for y axis}

\item{max_y}{Optional, set max value for y axis}

\item{colforcolor}{Optional, colored protein according to information in a column f the table}

\item{point_size}{Optional, allow for manually setting point sizes}

\item{label}{Set \code{FALSE} to remove label}

\item{label_size}{Optional, allow for manually setting label sizes}

\item{custom_text}{Optional, allow for setting custom text in labels}

\item{title}{Optional, allow for adding a title}
}
\value{
Return graph that is suitable for pdf export
}
\description{
\code{PDF_Plot} create a volcano plot using the ggplot2 package
}
\details{
If \code{colforcolor} is empty, protein with significantly different FC will be highlighted
If \code{colforcolor}, is filled, point will be colored according to information in the indicated column
}
