% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_test.R
\name{stat_test}
\alias{stat_test}
\title{Statistical analysis based on the glm model developped by EdgeR}
\usage{
stat_test(
  IPObj,
  control,
  treatment = NULL,
  batch = FALSE,
  div = c("nothing", "total", "DEseq"),
  filter = NULL,
  min.disp = NULL,
  glm = NULL
)
}
\arguments{
\item{IPObj}{object from the function \code{load_IP_Data}.}

\item{control}{denominator, name of the condition used as a reference. This name should correspond to one of the condition of the second
column of the \code{sampleTable}.}

\item{treatment}{Numerator, name of the condition to be compared with the \code{control}.
By default, the second condition found in the \code{sampleTable} is used as treatment.}

\item{batch}{Boolean-type parameter (default FALSE), if TRUE and if the \code{sampleTable} contains a third table
with batch indications, the batch effect are taken into account in the statistical model.}

\item{div}{Choice of the method of calculation for the scale factor}

\item{filter}{Supplemental filter. Minimal number of spectra to be considered as significant, indicated as a supplemental column with YES or NO}

\item{min.disp}{numeric scalar giving a value for the filtering out of low abundance tags for the calculation of the dispersion.
Only tags with total sum of counts above this value are used. Low abundance tags can adversely affect the dispersion estimation,
so this argument allows the user to select an appropriate filter threshold for the tag abundance. By defaut, this value is set to 5 by EdgeR.}

\item{glm}{Genewise Negative Binomial Generalized Linear Model with \code{"QL"} or without \code{"classic"} Quasi-likelihood tests.}
}
\value{
Return a dataframe with five colmumns: \code{logFC} (control used as denominator), \code{LR},
\code{p.value}, \code{adjp} and \code{number} corresponding the rank according to ajusted p-value.
The adjusted p-value is calculated according to the False Discovery Rate (FDR) method.
}
\description{
Function to perform the statistical analysis using the genewise negative binomial generalized linear model developped by the EdgeR package
By defaut, IPinquiry4 package uses GLM model with Quasi-likelihood Tests. By adding the argument `glm="classic"`,
you can use instead the EdgeR function based on the GLM without Quasi-likelihood Tests.
The \code{stat_test} function compares data from two conditions.
}
\details{
If \code{div} is \code{NULL} or \code{nothing}, scales factors are set to 1.
If \code{div} is \code{total}, scales factors are the total number of counts by sample.
If \code{div} is \code{DEseq}, scales factors are calculated according to the "median-to-ratio" method as in DEseq.
}
\examples{
CountTable <- system.file("extdata", "CountTable.txt", package = "IPinquiry4")
SampleTable <- system.file("extdata", "SampleTable.txt", package = "IPinquiry4")
IP_data <- load_IP_Data(CountTable, SampleTable)
test <- stat_test(IP_data, "urt1", treatment = "M1", div="DEseq")
}
