% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_IPObj_batch.R
\name{subset_IPObj_batch}
\alias{subset_IPObj_batch}
\title{Subset IPinquiry object according to batch}
\usage{
subset_IPObj_batch(IPObj, my_list = c())
}
\arguments{
\item{IPObj}{IPObj type object that comes directly from the function \code{load_IP_Data}}

\item{my_list}{List of the selected batchs (third column of the sampletable)}
}
\value{
IPObj object with only the selected batchs
}
\description{
Function to subset the IPObj according to the batch column of the SampleTable
}
