% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_IPObj_treat.R
\name{subset_IPObj_treat}
\alias{subset_IPObj_treat}
\title{Subset IPinquiry object according to treatment}
\usage{
subset_IPObj_treat(IPObj, my_list = c())
}
\arguments{
\item{IPObj}{object that comes directly from the function \code{load_IP_Data}}

\item{my_list}{List of the selected treatments (second column of the sampletable)}
}
\value{
IPObj object for the selected treatments
}
\description{
Function to subset the IPObj according to the treatment column of the SampleTable
}
