% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_counts.R
\name{total_counts}
\alias{total_counts}
\title{Fonction to count the total number of counts for each IP
Helene Zuber
\code{total_counts} This function allow for calculating the number of spectra for each IP
This could then be used as scale factor for statistical analysis}
\usage{
total_counts(data)
}
\arguments{
\item{data}{Objet de type IPObj, that comes from the function \code{loadData}.}
}
\value{
The function return the total number of counts by sample
}
\description{
Fonction to count the total number of counts for each IP
Helene Zuber
\code{total_counts} This function allow for calculating the number of spectra for each IP
This could then be used as scale factor for statistical analysis
}
